+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 <!--next-version-placeholder-->
 
+## v6.2.2 (2022-04-06)
+### Fix
+* Deletion email minor updates ([#1819](https://github.com/ocadotechnology/codeforlife-portal/issues/1819)) ([`4a8680d`](https://github.com/ocadotechnology/codeforlife-portal/commit/4a8680d8839bdf481ca2583e2806bb53f3a2cad0))
+
+## v6.2.1 (2022-04-05)
+### Fix
+* Fix sending email on account deletion ([#1818](https://github.com/ocadotechnology/codeforlife-portal/issues/1818)) ([`467b10a`](https://github.com/ocadotechnology/codeforlife-portal/commit/467b10aa665e883210fd856088323bb80cb74ad7))
+
+## v6.2.0 (2022-04-05)
+### Feature
+* Delete account function for teacher ([#1812](https://github.com/ocadotechnology/codeforlife-portal/issues/1812)) ([`0301952`](https://github.com/ocadotechnology/codeforlife-portal/commit/0301952cbf324a6267347f0b6c6d9a96b3ef62ff))
+
 ## v6.1.1 (2022-04-05)
 ### Fix
 * Hide anonymised teachers from dashboard ([#1816](https://github.com/ocadotechnology/codeforlife-portal/issues/1816)) ([`82b2dc3`](https://github.com/ocadotechnology/codeforlife-portal/commit/82b2dc33114c350abeeb2f1dea927152478872f2))